alert add

+++ Proyecto-Integrador/Integrador/src/services/products.js
@@ -1,5 +1,6 @@
 /* PRODUCTOS */
 
+import Swal from "sweetalert2";
 import { productoActivo } from "../../main.js";
 import {
 	handleGetProductLocalStorage,
@@ -40,6 +41,11 @@ const handleSaveOrModifyElements = () => {
 			categories,
 		};
 	}
+	Swal.fire({
+		title: "Correct!",
+		text: "Product Save!",
+		icon: "success",
+	});
 
 	setInLocalStorage(object);
 	handleGetProductsToStore();
@@ -49,11 +55,25 @@ const handleSaveOrModifyElements = () => {
 //Eliminar elemento
 
 export const handleDeleteProduct = () => {
-	const products = handleGetProductLocalStorage();
-	const result = products.filter((el) => el.id !== productoActivo.id);
-	// Setear el nuevo array
-	localStorage.setItem("products", JSON.stringify(result));
-	const newProducts = handleGetProductLocalStorage();
-	handleRenderList(newProducts);
-	closeModal();
+	Swal.fire({
+		title: "Are you sure?",
+		text: "You won't be able to revert this!",
+		icon: "warning",
+		showCancelButton: true,
+		confirmButtonColor: "#3085d6",
+		cancelButtonColor: "#d33",
+		confirmButtonText: "Yes, delete it!",
+	}).then((result) => {
+		if (result.isConfirmed) {
+			const products = handleGetProductLocalStorage();
+			const result = products.filter((el) => el.id !== productoActivo.id);
+			// Setear el nuevo array
+			localStorage.setItem("products", JSON.stringify(result));
+			const newProducts = handleGetProductLocalStorage();
+			handleRenderList(newProducts);
+			closeModal();
+		} else {
+			closeModal();
+		}
+	});
 };
